refactor : 패키지 구조 통일. lock 패키지 common 으로 변경
