--- src/main/java/com/lockers/outerpark/domain/concert/dto/RegisterConcertResponse.java
@@ -1,17 +1,0 @@
-package com.lockers.outerpark.domain.concert.dto;
-
-import java.time.LocalDateTime;
-
-import com.lockers.outerpark.domain.concert.entity.Concert;
-
-public record RegisterConcertResponse(String title, Long writerId, LocalDateTime createdAt, LocalDateTime updatedAt) {
-
-    public static RegisterConcertResponse of(Concert concert) {
-        return new RegisterConcertResponse(
-            concert.getTitle(),
-            concert.getWriter().getId(),
-            concert.getCreatedAt(),
-            concert.getUpdatedAt()
-        );
-    }
-}
