--- src/main/java/com/lockers/outerpark/domain/concert/dto/UpdateConcertResponse.java
@@ -1,17 +1,0 @@
-package com.lockers.outerpark.domain.concert.dto;
-
-import java.time.LocalDateTime;
-
-import com.lockers.outerpark.domain.concert.entity.Concert;
-
-public record UpdateConcertResponse(String title, Long writerId, LocalDateTime createdAt, LocalDateTime updatedAt) {
-
-    public static UpdateConcertResponse of(Concert concert) {
-        return new UpdateConcertResponse(
-            concert.getTitle(),
-            concert.getWriter().getId(),
-            concert.getCreatedAt(),
-            concert.getUpdatedAt()
-        );
-    }
-}
