+++ src/main/java/com/lockers/outerpark/domain/concert/dto/request/UpdateConcertRequest.java
@@ -1,0 +1,19 @@
+package com.lockers.outerpark.domain.concert.dto.request;
+
+import java.time.LocalDate;
+
+import lombok.Getter;
+
+@Getter
+public class UpdateConcertRequest {
+
+	private String title;
+
+	private Integer runningTime;
+
+	private Integer price;
+
+	private Integer limitAge;
+
+	private LocalDate performanceDate;
+}
