+++ src/main/java/com/lockers/outerpark/domain/concert/service/ConcertServiceImpl.java
@@ -5,11 +5,11 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.lockers.outerpark.domain.concert.dto.FindConcertResponse;
-import com.lockers.outerpark.domain.concert.dto.RegisterConcertRequest;
-import com.lockers.outerpark.domain.concert.dto.RegisterConcertResponse;
-import com.lockers.outerpark.domain.concert.dto.UpdateConcertRequest;
-import com.lockers.outerpark.domain.concert.dto.UpdateConcertResponse;
+import com.lockers.outerpark.domain.concert.dto.request.RegisterConcertRequest;
+import com.lockers.outerpark.domain.concert.dto.request.UpdateConcertRequest;
+import com.lockers.outerpark.domain.concert.dto.response.FindConcertResponse;
+import com.lockers.outerpark.domain.concert.dto.response.RegisterConcertResponse;
+import com.lockers.outerpark.domain.concert.dto.response.UpdateConcertResponse;
 import com.lockers.outerpark.domain.concert.entity.Concert;
 import com.lockers.outerpark.domain.concert.exception.ConcertErrorCode;
 import com.lockers.outerpark.domain.concert.exception.ConcertException;
