--- src/main/java/com/lockers/outerpark/domain/lock/exception/LockException.java
@@ -1,11 +1,0 @@
-package com.lockers.outerpark.domain.lock.exception;
-
-import com.lockers.outerpark.common.exception.BusinessException;
-
-public class LockException {
-    public static class SeatAlreadyLockedException extends BusinessException {
-        public SeatAlreadyLockedException() {
-            super(LockErrorCode.SEAT_LOCK_CONFLICT);
-        }
-    }
-}
